refactor: restructure application architecture and migrate to presentation layer
