+++ lib/main_staging.dart
@@ -1,5 +1,5 @@
-import 'package:ingenuity_flutter_app_template/app/app.dart';
 import 'package:ingenuity_flutter_app_template/bootstrap.dart';
+import 'package:ingenuity_flutter_app_template/presentation/app/app.dart';
 
 void main() {
   bootstrap(() => const App());
