+++ lib/presentation/app/view/app.dart
@@ -2,8 +2,8 @@ import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:flutter/material.dart';
 import 'package:ingenuity_flutter_app_template/amplifyconfiguration.dart';
-import 'package:ingenuity_flutter_app_template/counter/counter.dart';
 import 'package:ingenuity_flutter_app_template/l10n/l10n.dart';
+import 'package:ingenuity_flutter_app_template/presentation/counter/counter.dart';
 
 class App extends StatefulWidget {
   const App({super.key});
