+++ test/counter/cubit/counter_cubit_test.dart
@@ -1,7 +1,7 @@
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:ingenuity_flutter_app_template/counter/counter.dart';
+import 'package:ingenuity_flutter_app_template/presentation/counter/counter.dart';
 
 void main() {
   group('CounterCubit', () {
