+++ test/counter/view/counter_page_test.dart
@@ -2,7 +2,7 @@ import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:ingenuity_flutter_app_template/counter/counter.dart';
+import 'package:ingenuity_flutter_app_template/presentation/counter/counter.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../helpers/helpers.dart';
